[Glitch] Redesign: Remove old navigation in single-column mode
Port 9b7ba582c3dc56dbab1ce53fd71c51f810b56ebb to glitch-soc
