+++ app/javascript/flavours/glitch/features/navigation_panel/redesign/navigation_link.module.scss
@@ -61,6 +61,7 @@
 }
 
 .label {
+  max-width: 100%;
   min-width: 0;
   flex-grow: 1;
   flex-shrink: 1;
