+++ app/javascript/flavours/glitch/features/navigation_panel/redesign/styles.module.scss
@@ -5,6 +5,7 @@
   flex-direction: column;
   height: 100%;
   overflow-y: auto;
+  scrollbar-width: thin;
 }
 
 .list {
