+++ app/javascript/flavours/glitch/features/ui/components/columns_area/redesign.module.scss
@@ -1,0 +1,56 @@
+@use '@/styles/mastodon/variables' as *;
+
+.root {
+  --main-column-max-width: 640px;
+  --navigation-min-width: 280px;
+  --navigation-max-width: 360px;
+
+  box-sizing: border-box;
+  width: 100%;
+  height: 100%;
+  min-height: 100vh;
+  min-height: 100dvh;
+  padding-bottom: env(safe-area-inset-bottom);
+
+  @media (min-width: $no-gap-breakpoint) {
+    padding-inline: var(--space-2xs);
+  }
+
+  @media (min-width: $mobile-menu-breakpoint) {
+    display: grid;
+    grid-template-columns:
+      minmax(var(--navigation-min-width), 1fr)
+      minmax(0, var(--main-column-max-width))
+      minmax(0, 1fr);
+    gap: var(--space-xs);
+  }
+}
+
+.navigationWrapper {
+  position: sticky;
+  top: 0;
+  z-index: 1;
+  max-width: var(--navigation-max-width);
+  height: 100dvh;
+  padding-bottom: env(safe-area-inset-bottom);
+
+  @media (max-width: ($mobile-menu-breakpoint - 1)) {
+    display: none;
+  }
+}
+
+.main {
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  flex: 0 1 auto;
+  width: 100%;
+  contain: inline-size layout paint style;
+  container: column / inline-size;
+  color: var(--color-text-primary);
+  background-color: var(--color-bg-primary);
+
+  @media ($mobile-menu-breakpoint < width < $no-gap-breakpoint) {
+    border-inline: 1px solid var(--color-border-primary);
+  }
+}
